refactor: extract management panels from AgentWorkspace
